CREATED STYLES FOLDER - Moved .css files into Styles folder and got rid of extra folders inside Components folder.
